@@ -16,28 +16,27 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.1.0",
-    "@angular/common": "^7.1.0",
-    "@angular/compiler": "^7.1.0",
-    "@angular/core": "^7.1.0",
-    "@angular/forms": "^7.1.0",
-    "@angular/http": "^7.1.0",
-    "@angular/platform-browser": "^7.1.0",
-    "@angular/platform-browser-dynamic": "^7.1.0",
-    "@angular/platform-server": "^7.1.0",
-    "@angular/router": "^7.1.0",
+    "@angular/animations": "^8.2.14",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/platform-server": "^8.2.14",
+    "@angular/router": "^8.2.14",
     "@aspnet/signalr": "^1.1.4",
     "@hmcts/frontend": "^0.0.41-alpha",
-    "@nguniversal/module-map-ngfactory-loader": "7.1.0",
+    "@nguniversal/module-map-ngfactory-loader": "8.2.6",
     "@types/jquery": "^3.3.31",
-    "@types/webrtc": "0.0.25",
-    "adal-angular4": "^3.0.16",
+    "@types/webrtc": "^0.0.25",
+    "adal-angular4": "^4.0.12",
     "applicationinsights-js": "^1.0.20",
     "aspnet-prerendering": "^3.0.1",
     "bootstrap": "^4.4.1",
     "classlist.js": "^1.1.20150312",
-    "core-js": "^2.6.10",
-    "govuk-frontend": "^3.4.0",
+    "core-js": "^3.6.4",
+    "govuk-frontend": "^3.5.0",
     "guid-typescript": "^1.0.9",
     "jquery": "^3.4.1",
     "moment": "^2.24.0",
@@ -45,22 +44,22 @@
     "ngx-clipboard": "^12.3.0",
     "ngx-device-detector": "^1.3.20",
     "ngx-moment": "^3.5.0",
-    "popper.js": "^1.16.0",
-    "rxjs": "^6.5.3",
-    "webrtc-adapter": "^7.3.0",
-    "zone.js": "^0.8.29",
-    "snyk": "^1.259.0"
+    "@popperjs/core": "^2.0.0",
+    "rxjs": "^6.5.4",
+    "snyk": "^1.279.1",
+    "webrtc-adapter": "^7.4.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.9",
-    "@angular/cli": "^7.3.9",
-    "@angular/compiler-cli": "^7.2.15",
-    "@angular/language-service": "^7.1.0",
+    "@angular-devkit/build-angular": "~0.803.23",
+    "@angular/cli": "^8.3.23",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
     "@types/applicationinsights-js": "^1.0.9",
     "@types/jasmine": "^2.8.16",
     "@types/jasminewd2": "^2.0.8",
     "@types/node": "~8.10.59",
-    "codelyzer": "~5.2.1",
+    "codelyzer": "^5.0.1",
     "install-peers": "^1.0.3",
     "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
@@ -72,7 +71,7 @@
     "karma-junit-reporter": "^2.0.1",
     "node-sass": "^4.13.1",
     "puppeteer": "^2.0.0",
-    "typescript": "~3.1.1"
+    "typescript": "~3.5.3"
   },
   "optionalDependencies": {
     "protractor": "~5.4.0",
